Remove cmd refactoring and unit tests

+++ dvc/cmd_data_remove.py
@@ -1,10 +1,11 @@
 import os
-from boto.s3.connection import S3Connection
+
 import fasteners
+from boto.s3.connection import S3Connection
 
 from dvc.cmd_base import CmdBase
-from dvc.logger import Logger
 from dvc.exceptions import NeatLynxException
+from dvc.logger import Logger
 
 
 class DataRemoveError(NeatLynxException):
@@ -13,8 +14,8 @@ class DataRemoveError(NeatLynxException):
 
 
 class CmdDataRemove(CmdBase):
-    def __init__(self):
-        CmdBase.__init__(self)
+    def __init__(self, parse_config=True, git_obj=None, config_obj=None):
+        super(CmdDataRemove, self).__init__(parse_config, git_obj, config_obj)
         pass
 
     def define_args(self, parser):
@@ -22,66 +23,82 @@ class CmdDataRemove(CmdBase):
 
         parser.add_argument('target', metavar='', help='Target to remove - file or directory', nargs='*')
         parser.add_argument('-r', '--recursive', action='store_true', help='Remove directory recursively')
-        parser.add_argument('-c', '--remove-from-cloud', action='store_true', help='Keep file in cloud')
+        parser.add_argument('-l', '--keep-in-cloud', action='store_false', default=False,
+                            help='Do not remove data from cloud')
+        parser.add_argument('-c', '--keep-in-cache', action='store_false', default=False,
+                            help='Do not remove data from cache')
         pass
 
     def run(self):
         lock = fasteners.InterProcessLock(self.git.lock_file)
         gotten = lock.acquire(timeout=5)
         if not gotten:
-            Logger.printing('Cannot perform the command since DVC is busy and locked. Please retry the command later.')
+            Logger.info('Cannot perform the command since DVC is busy and locked. Please retry the command later.')
             return 1
 
         try:
-            if not self.skip_git_actions and not self.git.is_ready_to_go():
-                return 1
-
-            for target in self.args.target:
-                self.remove_target(target)
-
-            if self.skip_git_actions:
-                self.not_committed_changes_warning()
-                return 0
-
-            message = 'DVC data remove: {}'.format(' '.join(self.args.target))
-            self.git.commit_all_changes_and_log_status(message)
+            return self.remove_all_targets()
         finally:
             lock.release()
 
-    def remove_target(self, target):
-        if os.path.isdir(target):
-            if not self.args.recursive:
-                raise DataRemoveError('Directory cannot be removed. Use --recurcive flag.')
+    def remove_all_targets(self):
+        if not self.skip_git_actions and not self.git.is_ready_to_go():
+            return 1
 
-            if os.path.realpath(target) == \
-                    os.path.realpath(os.path.join(self.git.git_dir_abs, self.config.data_dir)):
-                raise DataRemoveError('data directory cannot be removed')
+        error = False
+        for target in self.args.target:
+            try:
+                if os.path.isdir(target):
+                    self.remove_dir(target)
+                else:
+                    self.remove_data_instance(target)
+            except NeatLynxException as ex:
+                Logger.error('Unable to remove data file "{}": {}'.format(target, ex))
+                error = True
 
-            return self.remove_dir(target)
+        message = 'DVC data remove: {}'.format(' '.join(self.args.target))
+        self.commit_if_needed(message, error)
 
-        data_path = self.path_factory.existing_data_path(target)
-        if os.path.islink(data_path.data.relative):
-            return self.remove_symlink(data_path)
+        return 0 if error == 0 else 1
 
-        raise DataRemoveError('Cannot remove a regular file "{}"'.format(target))
+    def remove_dir(self, target):
+        if not self.args.recursive:
+            raise DataRemoveError('Directory "%s" cannot be removed. Use --recurcive flag.' % target)
+
+        data_path = self.path_factory.data_path(target)
+        if data_path.data_dvc_short == '':
+            raise DataRemoveError('Data directory "%s" cannot be removed' % target)
+
+        return self.remove_dir_file_by_file(target)
 
     @staticmethod
     def remove_dir_if_empty(file):
         dir = os.path.dirname(file)
         if dir != '' and not os.listdir(dir):
             os.rmdir(dir)
+        pass
 
-    def remove_symlink(self, data_path):
-        if os.path.isfile(data_path.cache.relative):
+    def remove_data_instance(self, target):
+        # it raises exception if not a symlink is provided
+        data_path = self.path_factory.existing_data_path(target)
+
+        if not self.args.keep_in_cache and os.path.isfile(data_path.cache.relative):
             os.remove(data_path.cache.relative)
-            self.remove_dir_if_empty(data_path.cache.relavive)
+            self.remove_dir_if_empty(data_path.cache.relative)
+        else:
+            if not self.args.keep_in_cache:
+                Logger.warn(u'Unable to find cache file for data instance %s' % data_path.data.relative)
 
         if os.path.isfile(data_path.state.relative):
             os.remove(data_path.state.relative)
             self.remove_dir_if_empty(data_path.state.relative)
+        else:
+            Logger.warn(u'State file {} for data instance {} does not exist'.format(
+                data_path.state.relative, data_path.data.relative))
 
-        if self.args.remove_from_cloud:
-            self.remove_from_cloud(data_path.cache_file_aws_key)
+        if not self.args.keep_in_cloud:
+            aws_key = self.cache_file_aws_key(data_path.cache.dvc)
+            self.remove_from_cloud(aws_key)
 
         os.remove(data_path.data.relative)
         pass
@@ -96,19 +113,18 @@ class CmdDataRemove(CmdBase):
                 Logger.warn('S3 remove warning: file "{}" does not exist in S3'.format(aws_file_name))
             else:
                 key.delete()
-                Logger.printing('File "{}" was removed from S3'.format(aws_file_name))
+                Logger.info('File "{}" was removed from S3'.format(aws_file_name))
+        pass
 
-    def remove_dir(self, data_dir):
-        for f in os.listdir(data_dir):
-            fname = os.path.join(data_dir, f)
-            if os.path.isdir(fname):
-                self.remove_dir(fname)
-            elif os.path.islink(fname):
-                self.remove_symlink(fname)
+    def remove_dir_file_by_file(self, target):
+        for f in os.listdir(target):
+            file = os.path.join(target, f)
+            if os.path.isdir(file):
+                self.remove_dir_file_by_file(file)
             else:
-                raise DataRemoveError('Unsupported file type "{}"'.format(fname))
+                self.remove_data_instance(file)
 
-        os.rmdir(data_dir)
+        os.rmdir(target)
         pass
